% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_options.R
\name{add_options}
\alias{add_options}
\title{Add New Options to a Grammar}
\usage{
add_options(grammar, target_symbol, ...)
}
\arguments{
\item{grammar}{A data table containing a Tracery-like grammar}

\item{target_symbol}{A new or existing grammar symbol (character)}

\item{...}{A series of symbols (characters) to add to the target symbol}
}
\value{
A data table containing a Tracery-like grammar
}
\description{
\code{add_options} appends options to a new or existing symbol in a grammar.
}
