% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con2fcoales.R
\name{con2fcoales}
\alias{con2fcoales}
\title{Helper for coalescing vectors of different types}
\usage{
con2fcoales(u, z)
}
\arguments{
\item{u, z}{Vectors of equal length and of different data types (e.g. double and integer).
Missing values are accepted.}
}
\value{
A double data type vector of same length as the arguments.
}
\description{
This helper prevents the error in \link[data.table:coalesce]{fcoalesce}
when attempting to coalesce two vectors of different data type (double/integer).
}
