% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_story.R
\name{expand_story}
\alias{expand_story}
\title{Expand Story}
\usage{
expand_story(grammar, origin)
}
\arguments{
\item{grammar}{A data table containing a Tracery-like grammar}

\item{origin}{A specified start symbol to initiate story.}
}
\value{
A data table containing a Tracery-like grammar
}
\description{
Generate a story from a specified start symbol and grammar.
}
