% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_options.R
\name{remove_options}
\alias{remove_options}
\title{Remove Options from a Grammar}
\usage{
remove_options(grammar, target_symbol, ...)
}
\arguments{
\item{grammar}{A data table containing a Tracery-like grammar}

\item{target_symbol}{An existing grammar symbol (character)}

\item{...}{A series of symbols (characters) to remove from the target symbol}
}
\value{
A data table containing a Tracery-like grammar
}
\description{
\code{remove_options} removes options or an entire symbol from a grammar.
}
