% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_option.R
\name{select_option}
\alias{select_option}
\title{Select Option}
\usage{
select_option(grammar, target_symbol)
}
\arguments{
\item{grammar}{A data table containing a Tracery-like grammar}

\item{target_symbol}{An existing grammar symbol (character)}
}
\value{
A string (character)
}
\description{
Generate a pattern node by selection an option from a specified symbol.
}
