% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichDups.R
\name{whichDups}
\alias{whichDups}
\title{Find duplicated values in data}
\usage{
whichDups(x)
}
\arguments{
\item{x}{A data.table.}
}
\value{
A named character vector. Data columns containing distinct sets of duplicated
values have the names indexed.
}
\description{
The function finds duplicated values in each column of the data
file. Although not intended for direct use, it can be applied to a data file
once converted into "data.table" class.
}
\examples{

if (interactive()) {

  dir = system.file('extdata', package = 'replacer')
  setwd(dir)

  x = data.table::fread('data.csv', na.strings = c(NA_character_, ''))

  whichDups(x)

}

}
